+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 * Show breakpoint info as a tooltip if the breakpoint icon is clicked with
   LMB and held down.
 * Show stack as a hex dump, with options to view as short, int, long, ascii, ...
+* The "go to address" in the Assembly view now works if address it outside
+  current assembly view.
 
 ## [2.4] - 2024-03-18
 * Changed main icon to a more license friendly one.